+++ src/napari_imagej/widgets/menu.py
@@ -2,36 +2,28 @@
 The top-level menu for the napari-imagej widget.
 """
 from pathlib import Path
-from typing import Iterable, List, Optional
+from typing import Iterable, Optional
 
 from magicgui.widgets import request_values
 from napari import Viewer
 from napari._qt.qt_resources import QColoredSVGIcon
-from napari.layers import Image, Labels, Layer, Points, Shapes
+from napari.layers import Image, Layer
 from qtpy.QtCore import Qt, Signal
 from qtpy.QtGui import QIcon, QPixmap
-from qtpy.QtWidgets import (
-    QComboBox,
-    QDialog,
-    QDialogButtonBox,
-    QFrame,
-    QHBoxLayout,
-    QLabel,
-    QMessageBox,
-    QPushButton,
-    QVBoxLayout,
-    QWidget,
-)
+from qtpy.QtWidgets import QHBoxLayout, QMessageBox, QPushButton, QWidget
 from scyjava import is_arraylike
 
 from napari_imagej import settings
-from napari_imagej.java import ij, jc
+from napari_imagej.java import ij
 from napari_imagej.resources import resource_path
 from napari_imagej.utilities.event_subscribers import UIShownListener
 from napari_imagej.utilities.events import subscribe, unsubscribe
+from napari_imagej.widgets.widget_utils import _IMAGE_LAYER_TYPES, DetailExportDialog
 
 
 class NapariImageJMenu(QWidget):
+    """Container widget comprising the napari-imagej menu bar."""
+
     def __init__(self, viewer: Viewer):
         super().__init__()
         self.setLayout(QHBoxLayout())
@@ -43,8 +35,8 @@ class NapariImageJMenu(QWidget):
         self.to_ij: ToIJButton = ToIJButton(viewer)
         self.layout().addWidget(self.to_ij)
 
-        self.to_ij_extended: ToIJDetailedButton = ToIJDetailedButton(viewer)
-        self.layout().addWidget(self.to_ij_extended)
+        self.to_ij_detail: ToIJDetailedButton = ToIJDetailedButton(viewer)
+        self.layout().addWidget(self.to_ij_detail)
 
         self.gui_button: GUIButton = GUIButton(viewer)
         self.layout().addWidget(self.gui_button)
@@ -75,19 +67,23 @@ class NapariImageJMenu(QWidget):
     def finalize(self):
         # GUIButton initialization
         if not settings.headless():
-            self.gui_button._set_icon(resource_path("imagej2-16x16-flat"))
+            self.gui_button._icon_path = resource_path("imagej2-16x16-flat")
+            self.gui_button.setIcon(self.gui_button._icon())
             self.gui_button.setEnabled(True)
             self.gui_button.setToolTip("Display ImageJ2 UI")
-        # Subscribe UIShown subscriber
+        # Subscribe UIShownListener
         self.subscriber = UIShownListener()
         subscribe(ij(), self.subscriber)
 
     def __del__(self):
+        # Unsubscribe UIShownListener
         if self.subscriber:
             unsubscribe(ij(), self.subscriber)
 
 
 class IJMenuButton(QPushButton):
+    """Base class defining shared napari-imagej menu button functionality"""
+
     def __init__(self, viewer: Viewer):
         super().__init__()
         self.viewer = viewer
@@ -95,13 +91,26 @@ class IJMenuButton(QPushButton):
         self.recolor()
 
     def recolor(self, event=None):
-        if hasattr(self, "_icon"):
-            if isinstance(self._icon, QColoredSVGIcon):
-                self.setIcon(self._icon.colored(theme=self.viewer.theme))
+        """Recolors this button's icon, if it can be recolored."""
+        if icon := self._icon():
+            if isinstance(icon, QColoredSVGIcon):
+                self.setIcon(icon.colored(theme=self.viewer.theme))
+
+    def _icon(self):
+        """
+        Placeholder function returning this button's icon.
+        This function should be overwritten by subclasses.
+        """
+        return None
 
 
 class ToIJDetailedButton(IJMenuButton):
-    _icon = QColoredSVGIcon(resource_path("export_detailed"))
+    """
+    Button providing detailed image transfer from napari to ImageJ2.
+    Detailed image transfer allows the selection of an "image" layer
+    (either Image or Labels), along with the assignment of a "rois" layer
+    (either Points or Shapes), and the identification of dimension assignments
+    """
 
     def __init__(self, viewer: Viewer):
         super().__init__(viewer)
@@ -113,22 +122,11 @@ class ToIJDetailedButton(IJMenuButton):
         viewer.layers.selection.events.changed.connect(self.layer_selection_changed)
         self.clicked.connect(lambda: DetailExportDialog(self.viewer).exec())
 
-    def _handle_choices(self, choices):
-        # Queue UI call on the EDT
-        # TODO: Use EventQueue.invokeLater scyjava wrapper, once it exists
-        img = ij().convert().convert(ij().py.to_java(choices["image"]), jc.Dataset)
-        if choices["rois"]:
-            rois = ij().convert().convert(ij().py.to_java(choices["rois"]), jc.ROITree)
-            img.getProperties().put("rois", rois)
-        ij().ui().show(img)
-
-    def handle_no_choices(self):
-        RichTextPopup(
-            rich_message="There is no active window to export to ImageJ!",
-            exec=True,
-        )
+    def _icon(self):
+        return QColoredSVGIcon(resource_path("export_detailed"))
 
     def layer_selection_changed(self, event):
+        """Disables the button if there are no "image" layers"""
         for layer in self.viewer.layers:
             if isinstance(layer, _IMAGE_LAYER_TYPES):
                 self.setEnabled(True)
@@ -136,130 +134,7 @@ class ToIJDetailedButton(IJMenuButton):
         self.setEnabled(False)
 
 
-_IMAGE_LAYER_TYPES = (Image, Labels)
-_ROI_LAYER_TYPES = (Points, Shapes)
-
-
-class DetailExportDialog(QDialog):
-    def __init__(self, viewer: Viewer):
-        super().__init__()
-        self.setLayout(QVBoxLayout())
-        # Write the title to a Label
-        self.layout().addWidget(QLabel("Export data to ImageJ"))
-
-        # Parse layer options
-        self.imgs = []
-        self.rois = []
-        for layer in viewer.layers:
-            if isinstance(layer, _IMAGE_LAYER_TYPES):
-                self.imgs.append(layer)
-            elif isinstance(layer, _ROI_LAYER_TYPES):
-                self.rois.append(layer)
-
-        self.img_container = self.RichComboBox("Image:", self.imgs)
-        self.layout().addWidget(self.img_container)
-        self.roi_container = self.RichComboBox("ROIs:", self.rois, required=False)
-        self.layout().addWidget(self.roi_container)
-        self.dims_container = self.DimsComboBox(self.img_container)
-        self.layout().addWidget(self.dims_container)
-
-        current_layer: Layer = viewer.layers.selection.active
-        # The user likely wants to transfer the active layer, if it is an image
-        if isinstance(current_layer, _IMAGE_LAYER_TYPES):
-            self.img_container.combo.setCurrentText(current_layer.name)
-        else:
-            self.img_container.combo.setCurrentText(self.imgs[0].name)
-        self.dims_container.update(self.img_container.combo.currentIndex())
-        self.img_container.combo.currentIndexChanged.connect(self.dims_container.update)
-
-        self.buttons = QDialogButtonBox(QDialogButtonBox.Ok | QDialogButtonBox.Cancel)
-        self.layout().addWidget(self.buttons)
-        self.buttons.accepted.connect(self.accept)
-        self.buttons.rejected.connect(self.reject)
-
-    def accept(self):
-        super().accept()
-
-        def pass_to_ij():
-            # Grab the selected data
-            img = self.img_container.combo.currentData()
-            roi = self.roi_container.combo.currentData()
-
-            j_img = ij().py.to_java(img, dim_order=self.dims_container.selections())
-            if roi:
-                j_img.getProperties().put("rois", ij().py.to_java(roi))
-
-            ij().ui().show(j_img)
-
-        ij().thread().queue(lambda: pass_to_ij())
-
-    class RichComboBox(QWidget):
-        def __init__(self, title: str, choices: List[Layer], required=True):
-            super().__init__()
-            self.setLayout(QHBoxLayout())
-            self.layout().addWidget(QLabel(title))
-            self.combo = QComboBox()
-            self.choices = choices
-            if not required:
-                self.combo.addItem("--------", None)
-            for c in choices:
-                self.combo.addItem(c.name, c)
-            self.layout().addWidget(self.combo)
-
-    class DimsComboBox(QFrame):
-        def __init__(self, combo_box):
-            super().__init__()
-            self.selection_box = combo_box
-            self.setLayout(QVBoxLayout())
-            self.setFrameStyle(QFrame.Box)
-            self.layout().addWidget(QLabel("Dimensions:"))
-
-            self.dims = [
-                [],
-                ["X"],
-                ["Y", "X"],
-                ["Z", "Y", "X"],
-                ["T", "Y", "X", "C"],
-                ["T", "Z", "Y", "X", "C"],
-            ]
-
-        def update(self, index: int):
-            # remove old widgets
-            for child in self.children():
-                if isinstance(child, self.DimSelector):
-                    self.layout().removeWidget(child)
-                    child.deleteLater()
-            selected = self.selection_box.combo.itemData(index)
-            guess = self.dims[len(selected.data.shape)]
-            for i, g in enumerate(guess):
-                self.layout().addWidget(
-                    self.DimSelector(f"dim_{i} ({selected.data.shape[i]} px)", g)
-                )
-
-        def selections(self):
-            selections = []
-            for child in self.children():
-                if isinstance(child, self.DimSelector):
-                    selections.append(child.combo.currentText())
-            return selections
-
-        class DimSelector(QWidget):
-            choices = ["X", "Y", "Z", "T", "C", "Unspecified"]
-
-            def __init__(self, title: str, guess: str):
-                super().__init__()
-                self.setLayout(QHBoxLayout())
-                self.layout().addWidget(QLabel(title))
-                self.combo = QComboBox()
-                for c in self.choices:
-                    self.combo.addItem(c)
-                self.combo.setCurrentText(guess)
-                self.layout().addWidget(self.combo)
-
-
 class ToIJButton(IJMenuButton):
-    _icon = QColoredSVGIcon(resource_path("export"))
-
     def __init__(self, viewer: Viewer):
         super().__init__(viewer)
         self.setEnabled(False)
@@ -268,21 +143,17 @@ class ToIJButton(IJMenuButton):
         self.setToolTip("Export napari Layer")
         self.clicked.connect(self.send_active_layer)
 
-    def _set_icon(self, path: str):
-        icon: QIcon = QIcon(QPixmap(path))
-        self.setIcon(icon)
+    def _icon(self):
+        return QColoredSVGIcon(resource_path("export"))
 
     def send_active_layer(self):
-        active_layer: Optional[Layer] = self.viewer.layers.selection.active
-        if active_layer:
-            self._show(active_layer)
+        layer: Optional[Layer] = self.viewer.layers.selection.active
+        if layer:
+            # Queue UI call on the EDT
+            # TODO: Use EventQueue.invokeLater scyjava wrapper, once it exists
+            ij().thread().queue(lambda: ij().ui().show(ij().py.to_java(layer)))
         else:
             self.handle_no_choices()
-
-    def _show(self, layer):
-        # Queue UI call on the EDT
-        # TODO: Use EventQueue.invokeLater scyjava wrapper, once it exists
-        ij().thread().queue(lambda: ij().ui().show(ij().py.to_java(layer)))
 
     def handle_no_choices(self):
         RichTextPopup(
@@ -296,13 +167,14 @@ class ToIJButton(IJMenuButton):
 
 
 class FromIJButton(IJMenuButton):
-    _icon = QColoredSVGIcon(resource_path("import"))
-
     def __init__(self, viewer: Viewer):
         super().__init__(viewer)
 
         self.setToolTip("Import active ImageJ2 Dataset")
         self.clicked.connect(self.get_active_layer)
+
+    def _icon(self):
+        return QColoredSVGIcon(resource_path("import"))
 
     def _get_objects(self, t):
         compatibleInputs = ij().convert().getCompatibleInputs(t)
@@ -357,19 +229,20 @@ class FromIJButton(IJMenuButton):
 
 
 class GUIButton(IJMenuButton):
+    _icon_path = resource_path("imagej2-16x16-flat-disabled")
+
     def __init__(self, viewer: Viewer):
         super().__init__(viewer)
         self.setEnabled(False)
-        self._set_icon(resource_path("imagej2-16x16-flat-disabled"))
+        self.setIcon(self._icon())
 
         if settings.headless():
             self.setToolTip("ImageJ2 GUI unavailable!")
         else:
             self.setToolTip("Display ImageJ2 UI (loading)")
 
-    def _set_icon(self, path: str):
-        icon: QIcon = QIcon(QPixmap(path))
-        self.setIcon(icon)
+    def _icon(self):
+        return QIcon(QPixmap(self._icon_path))
 
     def disable_popup(self):
         RichTextPopup(
@@ -386,13 +259,14 @@ class SettingsButton(IJMenuButton):
     # Signal used to identify changes to user settings
     setting_change = Signal(bool)
 
-    _icon = QColoredSVGIcon(resource_path("gear"))
-
     def __init__(self, viewer: Viewer):
         super().__init__(viewer)
 
         self.clicked.connect(self._update_settings)
         self.setting_change.connect(self._handle_settings_change)
+
+    def _icon(self):
+        return QColoredSVGIcon(resource_path("gear"))
 
     def _update_settings(self):
         """
